fix: added dtype to h in smoothing ops (and dtype in tests)

+++ pylops/basicoperators/smoothing1d.py
@@ -87,7 +87,7 @@ class Smoothing1D(Convolve1D):
     ):
         if nsmooth % 2 == 0:
             nsmooth += 1
-        h = np.ones(nsmooth) / float(nsmooth)
+        h = np.ones(nsmooth, dtype=dtype) / float(nsmooth)
         offset = (nsmooth - 1) // 2
         super().__init__(
             dims=dims, h=h, axis=axis, offset=offset, dtype=dtype, name=name
